+++ stack-springboot-service/src/main/java/org/robinkyou/stack/demo/service/PersonService.java
@@ -5,5 +5,5 @@ import org.robinkyou.stack.demo.domain.common.BizResponse;
 import org.robinkyou.stack.demo.domain.po.Person;
 
 public interface PersonService {
-    BizResponse<Person> queryPersonByPersonId(BizRequest<Long> personId);
+    BizResponse<Person> queryPersonByPersonId(BizRequest<Integer> personId);
 }
